fix log_db_metrics
